move query-string-array-format.ts to enumarations folder
